+++ pytcl/containers/__init__.py
@@ -12,9 +12,48 @@ from pytcl.containers.kd_tree import (
     BallTree,
 )
 
+from pytcl.containers.rtree import (
+    BoundingBox,
+    merge_boxes,
+    box_from_point,
+    box_from_points,
+    RTreeNode,
+    RTreeResult,
+    RTree,
+)
+
+from pytcl.containers.vptree import (
+    VPTreeResult,
+    VPNode,
+    VPTree,
+)
+
+from pytcl.containers.covertree import (
+    CoverTreeResult,
+    CoverTreeNode,
+    CoverTree,
+)
+
 __all__ = [
+    # K-D Tree
     "KDNode",
     "NearestNeighborResult",
     "KDTree",
     "BallTree",
+    # R-Tree
+    "BoundingBox",
+    "merge_boxes",
+    "box_from_point",
+    "box_from_points",
+    "RTreeNode",
+    "RTreeResult",
+    "RTree",
+    # VP-Tree
+    "VPTreeResult",
+    "VPNode",
+    "VPTree",
+    # Cover Tree
+    "CoverTreeResult",
+    "CoverTreeNode",
+    "CoverTree",
 ]